+++ DelvUI/Interface/EnemyList/EnemyListConfig.cs
@@ -55,8 +55,8 @@ namespace DelvUI.Interface.EnemyList
         [NestedConfig("Health Label", 80)]
         public EditableLabelConfig HealthLabel = new EditableLabelConfig(new Vector2(30, 0), "[health:percent]%", DrawAnchor.Left, DrawAnchor.Left);
 
-        [NestedConfig("Order Letter Label", 90)]
-        public LabelConfig OrderLetterLabel = new LabelConfig(new Vector2(5, 0), "", DrawAnchor.Left, DrawAnchor.Left);
+        [NestedConfig("Order Label", 90)]
+        public DefaultFontLabelConfig OrderLabel = new DefaultFontLabelConfig(new Vector2(5, 0), "", DrawAnchor.Left, DrawAnchor.Left);
 
         [NestedConfig("Colors", 100)]
         public EnemyListHealthBarColorsConfig Colors = new EnemyListHealthBarColorsConfig();
@@ -79,7 +79,6 @@ namespace DelvUI.Interface.EnemyList
 
             config.NameLabel.FontID = FontsConfig.DefaultMediumFontKey;
             config.HealthLabel.FontID = FontsConfig.DefaultMediumFontKey;
-            config.OrderLetterLabel.FontID = FontsConfig.DefaultMediumFontKey;
 
             config.MouseoverAreaConfig.Enabled = false;